libwasi: rename wasi_poll.[ch] and wasi_path.[ch]
