+++ libwasi/CMakeLists.txt
@@ -3,8 +3,8 @@
 set(lib_wasi_sources
 	"wasi.c"
 	"wasi_fdtable.c"
-	"wasi_path.c"
-	"wasi_poll.c"
+	"wasi_path_subr.c"
+	"wasi_poll_subr.c"
 	"wasi_subr.c"
 	"wasi_host_pathop.c"
 	"wasi_host_fdop.c"
