+++ libwasi/wasi.c
@@ -76,8 +76,8 @@
 #include "wasi_host_pathop.h"
 #include "wasi_host_subr.h"
 #include "wasi_impl.h"
-#include "wasi_path.h"
-#include "wasi_poll.h"
+#include "wasi_path_subr.h"
+#include "wasi_poll_subr.h"
 #include "wasi_subr.h"
 #include "wasi_utimes.h"
 #include "xlog.h"
